renames module to _base
