+++ pgmpy/prediction/_base.py
